requirements.txt
-i https://pypi.org/simple
bleach==3.1.5
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
codecov==2.1.9
colorama==0.4.3
coverage==5.2.1
cryptography==3.1
decorator==4.4.2
docutils==0.16
flask-cors==3.0.8
flask==1.1.2
google-auth==1.20.1
idna==2.10
importlib-metadata==1.7.0 ; python_version < '3.8'
inflection==0.5.1
itsdangerous==1.1.0
jeepney==0.4.3 ; sys_platform == 'linux'
jinja2==2.11.2
k8-kat==0.0.156
keyring==21.3.1
kubernetes==10.0.1
markupsafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pkginfo==1.5.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
pyyaml==5.3.1
readme-renderer==26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
secretstorage==3.1.2 ; sys_platform == 'linux'
six==1.12.0
sqlalchemy==1.3.19
tqdm==4.48.2
twine==3.2.0
typing-extensions==3.7.4.3
urllib3==1.25.10
validators==0.18.0
webencodings==0.5.1
websocket-client==0.57.0
werkzeug==1.0.1
zipp==3.1.0
